Enable all platform tests (esp CentOS 7,8)
Includes:

- adjust all of the Dockerfile templates to be the same, even for CentOS 8.
- adjust molecule config to enable systemd on CentOS containers
- enable nearly all platforms in all tests
- add Travis timeout workaround (see below)

Travis times out if no output is received for 10 min. But, the
idempotence check stage re-runs the converge playbook and swallows the
output. As the converge stage can take more than 10 min, we needed a way
to let travis know not to kill the job. So, we print a `.` ever 5 min
from a background process.

This travis fix should make the builds more reliable, even though we're
enabling extra platforms to test in this commit.