Add create_for_localhost_exception parameter to mongodb_user module (#182)
* exclude built collections

* update mongodb_user doc

* Add create_for_localhost_exception parameter to mongodb_user module

Simplify touching a file when creating an admin user with the localhost
exception. A pain point with the localhost exception is that you can't
tell if the first admin account has been added without attempting to
authenticate. So, provide a convenient helper that will touch the given
path when the user is successfully added.

Note that the name is long and unwieldy on purpose. This should be a
relatively uncommon use-case when adding or updating most users. So, we
use the length as a signal to users that "this is not the parameter you
were looking for".

Also clarifies comment why ignore Unauthorized: Whether or not the
localhost exception applies, we should still ignore aunauthorized while
searching for an existing user.

* begin investigating integration tests for mongodb_user

* consolidate mongod stop logic in mongodb_user int test

* mongodb_user int tests: refactor replicaset config/start tasks

* add integration test for create_for_localhost_exception

* fix create_for_localhost_exception file skip

* fix whitespace