Document roles & plugins in the readme (#343)
* Document roles & plugins in the readme

If you're new to setting up mongodb, figuring out which role to use can be
confusing. The `mongodb_config` role is especially problematic, because
standalone installations do not have a `config` server. The only "config" for
standalone mongodb instances is the mongo.conf file, which is completely
different. So, this adds a little map to help guide people in the right
direction.

* Use FQCN in README

* Clarify role descriptions

* update role description in role metadata

* add description to galaxy.yml