Pass login credentials when instantiating new MongoClient
In plugins/modules/mongodb_user.py, we have the following function call:

client = get_mongodb_client(module, directConnection=directConnection)

when it should be

client = get_mongodb_client(module, login_user, login_password, login_database, directConnection=directConnection)

Then, the values of the username as password needed to authenticate in the database are passed to MongoClient() as None values. As a consequence, the connection to the database fails.

This results in the error message, "Unable to connect to database: SCRAM-SHA-256 requires a username".

To make the function more robust, grab the login_user, login_password, and login_database from module.